requirements-ml.txt
numpy>=1.24
pandas>=2.0
scikit-learn>=1.3
psycopg2-binary>=2.9
